+++ CHANGELOG.md
@@ -8,4 +8,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Asserts:
-  - `Assert-BlockString`
+  - `Assert-BlockString` (alias `Should-BeBlockString`)